fix possible compiler warnings
changed pointer cast to uintptr_t in tlsf
unsigned long might be only 32 bits long

+++ core/ov/source/libov/tlsf.c
@@ -521,7 +521,7 @@ size_t init_memory_pool(size_t mem_pool_size, void *mem_pool)
         return -1;
     }
 
-    if (((unsigned long) mem_pool & PTR_MASK)) {
+    if (((uintptr_t) mem_pool & PTR_MASK)) {
         ERROR_MSG("init_memory_pool (): mem_pool must be aligned to a word\n");
         return -1;
     }
@@ -584,7 +584,7 @@ size_t add_new_area(void *area, size_t area_size, void *mem_pool)
         lb1 = ptr->end;
 
         /* Merging the new area with the next physically contigous one */
-        if ((unsigned long) ib1 == (unsigned long) lb0 + BHDR_OVERHEAD) {
+        if ((uintptr_t) ib1 == (uintptr_t) lb0 + BHDR_OVERHEAD) {
             if (tlsf->area_head == ptr) {
                 tlsf->area_head = ptr->next;
                 ptr = ptr->next;
@@ -605,7 +605,7 @@ size_t add_new_area(void *area, size_t area_size, void *mem_pool)
 
         /* Merging the new area with the previous physically contigous
            one */
-        if ((unsigned long) lb1->ptr.buffer == (unsigned long) ib0) {
+        if ((uintptr_t) lb1->ptr.buffer == (uintptr_t) ib0) {
             if (tlsf->area_head == ptr) {
                 tlsf->area_head = ptr->next;
                 ptr = ptr->next;
@@ -1373,8 +1373,8 @@ int tlsf_check_pool(void* _tlsf){
 void dump_memory_region(unsigned char *mem_ptr, unsigned int size)
 {
 
-    unsigned long begin = (unsigned long) mem_ptr;
-    unsigned long end = (unsigned long) mem_ptr + size;
+	uintptr_t begin = (uintptr_t) mem_ptr;
+	uintptr_t end = (uintptr_t) mem_ptr + size;
     int column = 0;
 
     begin >>= 2;
